pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "notebooks"]

[project]
name = "src"
version = "0.1.0"
description = "Exploratory analysis of the ade20k dataset"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "canvit-probes",
    "canvit-pytorch",
    "dinov3",
    "torch>=2.0",
    "torchvision",
    "matplotlib",
    "pillow",
    "pandas",
    "numpy",
    "tqdm",
    "jupytext>=1.19.1",
    "ipykernel>=7.2.0",
    "pypatree>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.basedpyright]
pythonVersion = "3.13"
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
testpaths = ["test"]

[tool.uv.sources]
dinov3 = { git = "https://github.com/yberreby/dinov3.git" }
canvit-pytorch = { git = "https://github.com/m2b3/CanViT-PyTorch.git" }
canvit-probes = { git = "ssh://git@github.com/m2b3/CanViT-probes.git", branch = "main" }

[dependency-groups]
dev = [
    "basedpyright>=1.37.1",
    "ipython>=9.9.0",
    "pytest>=8.0.0",
    "ruff>=0.14.12",
]
